Tests/LibWeb/Text/input/css/invalid-font-variant-css-text.html

<!DOCTYPE html>
<html>
    <style>
        #foo {
            font-variant: normal;
            font-variant-ligatures: none;
            font-variant-caps: all-petite-caps;
        }
    </style>
    <script src="../include.js"></script>
    <script>
        test(() => {
            println(document.styleSheets[0].cssRules[0].cssText);
        });
    </script>
</html>
